admin templates, duplicate templates, new organization shows on reload, update admin navbar
