+++ app/admin/template-list/page.tsx
@@ -1,0 +1,8 @@
+"use client";
+
+import NavBar from "@/components/NavBar/NavBar";
+import TemplateListPage from "@/components/TemplateList/TemplateList";
+
+export default function AdminTemplateListPage() {
+  return <TemplateListPage navBar={<NavBar />} showSidebar={false} />;
+}
